refactor components structure
