+++ src/components/bars/index.tsx
@@ -1,0 +1,2 @@
+export { default as NavBar } from './NavBar';
+export { default as SideBar } from './SideBar';
