+++ src/components/cards/index.ts
@@ -1,0 +1,3 @@
+export { default as Card } from './Card';
+export { default as CollateralsCard } from './CollateralsCard';
+export { default as ItemCard } from './ItemCard';
