+++ src/components/filters/index.tsx
@@ -1,0 +1,2 @@
+export { default as FilterTagPanel } from './FilterTagPanel';
+export { default as FilterTag } from './FilterTag';
