+++ src/components/index.ts
@@ -1,4 +1,19 @@
-export { default as ResumeData } from './ResumeData';
-export { default as MainDAICard } from './MainDAICard';
-export { default as SideBar } from './SideBar';
-export { default as NavBar } from './NavBar';
+export { NavBar, SideBar } from './bars';
+export { SpellList } from './lists/SpellList';
+export { CollateralListContainer } from './lists';
+export { CollateralsStructureErrorBoundary } from './errors';
+export { FilterTagPanel, FilterTag } from './filters';
+export { Icon, getIconByAsset } from './Icon';
+export { PageWrapper, MainWrapper } from './wrappers';
+export type { IconNames } from './Icon';
+export { ThemeProvider } from './styles';
+export { Card, CollateralsCard, ItemCard } from './cards';
+export { DatePicker, Input, Select } from './inputs';
+export {
+  Flex,
+  Label,
+  Spinner,
+  Table,
+  HorizontalLine,
+  VerticalLine,
+} from './styledComponents';
