+++ src/components/inputs/index.ts
@@ -1,0 +1,3 @@
+export { default as Input } from './Input';
+export { default as DatePicker } from './DatePicker';
+export { default as Select } from './Select';
