+++ src/components/lists/CollateralList/index.ts
@@ -1,2 +1,2 @@
+// eslint-disable-next-line import/prefer-default-export
 export { default as CollateralListContainer } from './CollateralListContainer';
-export { default as CollateralList } from './CollateralList';
