+++ src/components/lists/SpellList/cells.tsx
@@ -3,8 +3,8 @@
 import React, { useLayoutEffect, useRef } from 'react';
 import styled from 'styled-components';
 import { getColorFromStatus } from '../../../services/utils/color';
-import { getEtherscanAddressLinkFromHash } from '../../../services/utils/links';
 import { formatDate, isPlural } from '../../../services/utils/formatsFunctions';
+import { getEtherscanAddressLinkFromHash } from '../../../services/utils/links';
 
 export const LabelCell = ({
   color,
