+++ src/components/lists/SpellList/changeColumns.tsx
