+++ src/components/lists/SpellList/spellColumns.tsx
