+++ src/components/lists/index.ts
@@ -1,0 +1,2 @@
+export { SpellList } from './SpellList';
+export { CollateralListContainer } from './CollateralList';
