+++ src/components/styledComponents/index.ts
@@ -1,11 +1,6 @@
 export { default as Flex } from './Flex';
 export { default as HorizontalLine } from './HorizontalLine';
 export { default as VerticalLine } from './VerticalLine';
-export { default as Card } from './Card';
-export { default as FilterTag } from './FilterTag';
 export { default as Spinner } from './Spinner';
-export { SpellList } from './SpellList';
-export { default as CollateralsCard } from './CollateralsCard';
 export { default as Label } from './Label';
-export { default as Input } from './Input';
-export { default as Select } from './Select';
+export { default as Table } from './Table';
