+++ src/components/wrappers/index.ts
@@ -1,0 +1,3 @@
+/* eslint-disable import/prefer-default-export */
+export { default as PageWrapper } from './PageWrapper';
+export { default as MainWrapper } from './MainWrapper';
